退出

+++ js/login.js
@@ -2,13 +2,13 @@
  * @Author: 吴世强 11076535+mr-wushiqiang@user.noreply.gitee.com
  * @Date: 2022-08-31 15:34:12
  * @LastEditors: 吴世强 11076535+mr-wushiqiang@user.noreply.gitee.com
- * @LastEditTime: 2022-08-31 16:03:36
+ * @LastEditTime: 2022-08-31 16:08:44
  * @FilePath: \wsq_demo01\js\login.js
  * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
  */
 document.write(`<script src="https://cdn.bootcss.com/blueimp-md5/2.10.0/js/md5.min.js"></script>`);
 
-function check() {
+function login() {
   var userName = document.getElementById('username');
   var pwd = document.getElementById('password');
   console.log(pwd.value);
@@ -37,6 +37,8 @@ function check() {
 }
 
 function logout() {
-  window.location.replace('https://51930236.github.io/wsq_demo01/login.html');\
-  localStorage.clear()
+  if (localStorage.getItem("CURRENT_LOGIN_INFO")) {
+    localStorage.clear()
+    window.location.replace('https://51930236.github.io/wsq_demo01/login.html');
+  }
 }
